test(dedupe): consolidate execution_telemetry tests (5 → 3 files)
Consolidate ExecutionTelemetry test files in tests/unit/gpt_trader/tui/services:
- test_execution_telemetry_collector_core.py + submission_record.py → test_execution_telemetry.py
- test_execution_telemetry_execution_metrics.py + issue_tracking.py → test_execution_telemetry_metrics.py
- test_execution_telemetry_reason_tracking.py → test_execution_telemetry_reason.py (rename)

Cluster: 68f5edc7f33a
Files: 5 → 3 (-2)
Tests: 37 passing
