+++ tests/_triage/dedupe_candidates.yaml
@@ -2433,7 +2433,9 @@ clusters:
     expected_file_delta: -4
     priority: low
     rationale: Consolidate 5 test files into single parametrized file
-    status: pending
+    status: done
+    owner: claude
+    pr_url: https://github.com/Solders-Girdles/GPT-Trader/pull/313
     added: '2026-01-20'
   7ae3ae24a668:
     type: similar_names
