+++ tests/unit/gpt_trader/tui/services/test_execution_telemetry_reason.py
